+++ crates/gen/src/helpers.rs
@@ -4,60 +4,6 @@
 
 use crate::*;
 
-pub fn gen_type_name(def: &TypeDef, gen: &Gen) -> TokenStream {
-    format_name(def, gen, to_ident, false)
-}
-
-pub fn gen_abi_name(def: &TypeDef, gen: &Gen) -> TokenStream {
-    format_name(def, gen, to_abi_ident, false)
-}
-
-pub fn gen_turbo_abi_name(def: &TypeDef, gen: &Gen) -> TokenStream {
-    format_name(def, gen, to_abi_ident, true)
-}
-
-fn format_name<F>(def: &TypeDef, gen: &Gen, format_name: F, turbo: bool) -> TokenStream
-where
-    F: FnOnce(&str) -> TokenStream,
-{
-    let type_name = def.type_name();
-
-    if type_name.namespace.is_empty() {
-        let name = format_name(&scoped_name(def));
-        quote! { #name }
-    } else {
-        let namespace = gen.namespace(type_name.namespace);
-        let name = format_name(type_name.name);
-
-        if def.generics.is_empty() {
-            quote! { #namespace#name }
-        } else {
-            let colon_separated = if turbo || !namespace.as_str().is_empty() {
-                quote! { :: }
-            } else {
-                quote! {}
-            };
-
-            let generics = def.generics.iter().map(|g| gen_name(g, gen));
-            quote! { #namespace#name#colon_separated<#(#generics),*> }
-        }
-    }
-}
-
-fn scoped_name(def: &TypeDef) -> String {
-    if let Some(enclosing_type) = def.enclosing_type() {
-        if let Some(nested_types) = enclosing_type.nested_types() {
-            for (index, (nested_type, _)) in nested_types.iter().enumerate() {
-                if *nested_type == def.name() {
-                    return format!("{}_{}", &scoped_name(&enclosing_type), index);
-                }
-            }
-        }
-    }
-
-    def.name().to_string()
-}
-
 pub fn gen_type(def: &TypeDef, gen: &Gen, include: TypeInclude) -> TokenStream {
     match def.kind() {
         TypeKind::Interface => {
@@ -80,20 +26,6 @@ pub fn gen_type(def: &TypeDef, gen: &Gen, include: TypeInclude) -> TokenStream {
     }
 }
 
-pub fn gen_abi_type(def: &TypeDef, gen: &Gen) -> TokenStream {
-    match def.kind() {
-        TypeKind::Enum => gen_type_name(def, gen),
-        TypeKind::Struct => {
-            if def.is_blittable() {
-                gen_type_name(def, gen)
-            } else {
-                gen_abi_name(def, gen)
-            }
-        }
-        _ => quote! { ::windows::RawPtr },
-    }
-}
-
 pub fn gen_phantoms(def: &TypeDef) -> impl Iterator<Item = TokenStream> + '_ {
     def.generics.iter().map(move |g| {
         let g = gen_name(g, &Gen::Absolute);
@@ -111,6 +43,31 @@ pub fn gen_constraints(def: &TypeDef) -> TokenStream {
         .collect()
 }
 
+pub fn gen_winrt_constraint(sig: &MethodSignature, gen: &Gen) -> TokenStream {
+    let params = sig.params.iter().map(|p| gen_winrt_produce_type(p, gen));
+
+    let return_type = if let Some(return_type) = &sig.return_type {
+        let tokens = gen_name(&return_type.kind, gen);
+
+        if return_type.is_array {
+            quote! { ::windows::Array<#tokens> }
+        } else {
+            tokens
+        }
+    } else {
+        quote! { () }
+    };
+
+    quote! { F: FnMut(#(#params),*) -> ::windows::Result<#return_type> + 'static }
+}
+
+pub fn gen_method_constraints(params: &[MethodParam]) -> TokenStream {
+    if params.iter().any(|param| param.is_convertible()) {
+        quote! { 'a, }
+    } else {
+        quote! {}
+    }
+}
 
 pub fn gen_default(def: &ElementType) -> TokenStream {
     match def {
@@ -136,142 +93,7 @@ pub fn gen_default(def: &ElementType) -> TokenStream {
     }
 }
 
-pub fn gen_name(def: &ElementType, gen: &Gen) -> TokenStream {
-    match def {
-        ElementType::Void => quote! { ::std::ffi::c_void },
-        ElementType::Bool => quote! { bool },
-        ElementType::Char => quote! { u16 },
-        ElementType::I8 => quote! { i8 },
-        ElementType::U8 => quote! { u8 },
-        ElementType::I16 => quote! { i16 },
-        ElementType::U16 => quote! { u16 },
-        ElementType::I32 => quote! { i32 },
-        ElementType::U32 => quote! { u32 },
-        ElementType::I64 => quote! { i64 },
-        ElementType::U64 => quote! { u64 },
-        ElementType::F32 => quote! { f32 },
-        ElementType::F64 => quote! { f64 },
-        ElementType::ISize => quote! { isize },
-        ElementType::USize => quote! { usize },
-        ElementType::String => {
-            quote! { ::windows::HSTRING }
-        }
-        ElementType::IInspectable => {
-            quote! { ::windows::IInspectable }
-        }
-        ElementType::Guid => {
-            quote! { ::windows::Guid }
-        }
-        ElementType::IUnknown => {
-            quote! { ::windows::IUnknown }
-        }
-        ElementType::HRESULT => {
-            quote! { ::windows::HRESULT }
-        }
-        ElementType::Array((kind, len)) => {
-            // TODO: rename all Signature vars to "sig"
-            let name = gen_sig(kind, gen);
-            let len = Literal::u32_unsuffixed(*len);
-            quote! { [#name; #len] }
-        }
-        ElementType::GenericParam(generic) => generic.into(),
-        ElementType::MethodDef(def) => def.name().into(),
-        ElementType::Field(field) => field.name().into(),
-        ElementType::TypeDef(t) => gen_type_name(t, gen),
-        _ => unimplemented!(),
-    }
-}
 
-pub fn gen_abi_type_name(def: &ElementType, gen: &Gen) -> TokenStream {
-    match def {
-        ElementType::Void => quote! { ::std::ffi::c_void },
-        ElementType::Bool => quote! { bool },
-        ElementType::Char => quote! { u16 },
-        ElementType::I8 => quote! { i8 },
-        ElementType::U8 => quote! { u8 },
-        ElementType::I16 => quote! { i16 },
-        ElementType::U16 => quote! { u16 },
-        ElementType::I32 => quote! { i32 },
-        ElementType::U32 => quote! { u32 },
-        ElementType::I64 => quote! { i64 },
-        ElementType::U64 => quote! { u64 },
-        ElementType::F32 => quote! { f32 },
-        ElementType::F64 => quote! { f64 },
-        ElementType::ISize => quote! { isize },
-        ElementType::USize => quote! { usize },
-        ElementType::String => {
-            quote! { ::windows::RawPtr }
-        }
-        ElementType::IInspectable => {
-            quote! { ::windows::RawPtr }
-        }
-        ElementType::Guid => {
-            quote! { ::windows::Guid }
-        }
-        ElementType::IUnknown => {
-            quote! { ::windows::RawPtr }
-        }
-        ElementType::HRESULT => {
-            quote! { ::windows::HRESULT }
-        }
-        ElementType::Array((kind, len)) => {
-            let name = gen_abi_sig(kind, gen);
-            let len = Literal::u32_unsuffixed(*len);
-            quote! { [#name; #len] }
-        }
-        ElementType::GenericParam(generic) => {
-            let name = format_token!("{}", generic);
-            quote! { <#name as ::windows::Abi>::Abi }
-        }
-        ElementType::TypeDef(def) => gen_abi_type(def, gen),
-        _ => unimplemented!(),
-    }
-}
-
-fn gen_namespaces<'a>(
-    namespaces: &'a BTreeMap<&'static str, TypeTree>,
-) -> impl Iterator<Item = TokenStream> + 'a {
-    namespaces.iter().map(move |(name, tree)| {
-        if tree.include {
-            let name = to_ident(name);
-
-            let tokens = gen_tree(tree);
-
-            quote! {
-                // TODO: https://github.com/microsoft/windows-rs/issues/212
-                // TODO: https://github.com/microsoft/win32metadata/issues/380
-                #[allow(unused_variables, non_upper_case_globals, non_snake_case, unused_unsafe, non_camel_case_types, dead_code, clippy::all)]
-                pub mod #name {
-                    #(#tokens)*
-                }
-            }
-         } else {
-             TokenStream::new()
-         }
-    })
-}
-
-pub fn gen_tree(tree: &TypeTree) -> impl Iterator<Item = TokenStream> + '_ {
-    let gen = Gen::Relative(tree.namespace);
-
-    tree.types
-        .iter()
-        .map(move |t| gen_type_entry(t.1, &gen))
-        .chain(gen_namespaces(&tree.namespaces))
-}
-
-pub fn gen_type_entry(entry: &TypeEntry, gen: &Gen) -> TokenStream {
-    if entry.include == TypeInclude::None {
-        return TokenStream::new();
-    }
-
-    match &entry.def {
-        ElementType::TypeDef(def) => gen_type(&def.clone().with_generics(), gen, entry.include),
-        ElementType::MethodDef(def) => gen_function(def, gen),
-        ElementType::Field(def) => gen_constant(def, gen),
-        _ => unimplemented!(),
-    }
-}
 
 pub fn gen_sig(sig: &Signature, gen: &Gen) -> TokenStream {
     let mut tokens = TokenStream::new();
@@ -322,24 +144,6 @@ pub fn gen_sig_default(sig: &Signature) -> TokenStream {
     }
 }
 
-pub fn gen_winrt_constraint(sig: &MethodSignature, gen: &Gen) -> TokenStream {
-    let params = sig.params.iter().map(|p| gen_winrt_produce_type(p, gen));
-
-    let return_type = if let Some(return_type) = &sig.return_type {
-        let tokens = gen_name(&return_type.kind, gen);
-
-        if return_type.is_array {
-            quote! { ::windows::Array<#tokens> }
-        } else {
-            tokens
-        }
-    } else {
-        quote! { () }
-    };
-
-    quote! { F: FnMut(#(#params),*) -> ::windows::Result<#return_type> + 'static }
-}
-
 pub fn gen_win32_abi(sig: &MethodSignature, gen: &Gen) -> TokenStream {
     // TODO: param insead of p consistency
     let params = sig.params.iter().map(|p| {
@@ -376,7 +180,8 @@ pub fn gen_winrt_abi(sig: &MethodSignature, gen: &Gen) -> TokenStream {
             let abi = gen_abi_sig(&p.signature, gen);
 
             if p.signature.is_array {
-                let abi_size_name = gen_param_abi_size_name(&p.param);
+                let abi_size_name = to_ident(&format!("{}_array_size", p.param.name()));
+
                 if p.param.is_input() {
                     quote! { #abi_size_name: u32, #name: *const #abi }
                 } else if p.signature.by_ref {
@@ -507,7 +312,15 @@ pub fn gen_winrt_method(
             &sig.params
         };
 
-    let name = gen_method_info_name(sig, method, interface);
+    let name =     if (interface.kind == InterfaceKind::Composable || interface.kind == InterfaceKind::Extend)
+    && sig.params.len() == 2
+{
+    "new".into()
+} else if method.overload > 1 {
+    format_token!("{}{}", &method.name, method.overload)
+} else {
+    method.name.clone().into()
+};
 
     let vtable_offset = Literal::u32_unsuffixed(method.vtable_offset);
     let constraints = gen_method_constraints(params);
@@ -616,30 +429,6 @@ pub fn gen_winrt_method(
                 }
             }
         }
-    }
-}
-
-fn gen_method_info_name(
-    sig: &MethodSignature,
-    method: &MethodInfo,
-    interface: &InterfaceInfo,
-) -> TokenStream {
-    if (interface.kind == InterfaceKind::Composable || interface.kind == InterfaceKind::Extend)
-        && sig.params.len() == 2
-    {
-        "new".into()
-    } else if method.overload > 1 {
-        format_token!("{}{}", &method.name, method.overload)
-    } else {
-        method.name.clone().into()
-    }
-}
-
-pub fn gen_method_constraints(params: &[MethodParam]) -> TokenStream {
-    if params.iter().any(|param| param.is_convertible()) {
-        quote! { 'a, }
-    } else {
-        quote! {}
     }
 }
 
@@ -855,12 +644,3 @@ pub fn gen_winrt_upcall(sig: &MethodSignature, inner: TokenStream, gen: &Gen) ->
         },
     }
 }
-
-pub fn gen_param_name(param: &Param) -> TokenStream {
-    to_ident(&param.name().to_lowercase())
-}
-
-pub fn gen_param_abi_size_name(param: &Param) -> TokenStream {
-    to_ident(&format!("{}_array_size", param.name()))
-}
-
